Move methods to helper folder
