+++ src/js/AuthButton.jsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { Link, withRouter } from 'react-router-dom';
 import { Button } from 'react-bootstrap';
-import appAuth from './appAuth';
+import appAuth from './helpers/appAuth';
 
 const AuthButton = withRouter(
   ({ history }) =>
