+++ src/js/helpers/appAuth.js
@@ -1,4 +1,4 @@
-import { auth } from '../firebase';
+import { auth } from '../../firebase';
 
 const appAuth = {
   isAuthenticated: false,
