+++ src/js/helpers/renderMergedProps.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const renderMergedProps = (component, ...rest) => {
+  const finalProps = Object.assign({}, ...rest);
+  return React.createElement(component, finalProps);
+};
+
+export default renderMergedProps;
